+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+Unreleased
+----------
+
+- v3: bump openapi schema - thanos + kube audit
+
 3.1.25
 ------